+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 This log documents significant changes for each release.  This project follows
 [Semantic Versioning](http://semver.org/).
 
+## [18.3.0] - 2021-12-14
+### Changed
+- Added request for extra items when some are selected to keep the number of
+  dropdown items.
+
 ## [18.2.0] - 2021-12-07
 ### Changed
 - Added options.search for specifying a function to return results.